Add new package pilz_utils
- move utilities into new package
- add new package to coverage build
