+++ .travis.yml
@@ -25,7 +25,7 @@ env:
       CATKIN_LINT=false NOT_TEST_BUILD=true
       AFTER_SCRIPT="./.coverage.sh"
       CMAKE_ARGS="-DENABLE_COVERAGE_TESTING=ON -DCMAKE_BUILD_TYPE=Debug"
-      DOCKER_RUN_OPTS='-e COVERAGE_PKGS="prbt_hardware_support"'
+      DOCKER_RUN_OPTS='-e COVERAGE_PKGS="prbt_hardware_support pilz_utils"'
 matrix:
   allow_failures:
     - env: ROS_REPO=ros
