+++ pilz_utils/README.md
@@ -1,0 +1,3 @@
+# pilz_utils
+
+This package provides some tools used by Pilz packages for easy re-usability.
