+++ pilz_utils/test/unittest_get_param.cpp
@@ -1,18 +1,17 @@
 /*
- * Copyright (c) 2018 Pilz GmbH & Co. KG
+ * Copyright (c) 2019 Pilz GmbH & Co. KG
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
-
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
-
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 #include <string>
@@ -21,9 +20,9 @@
 
 #include <ros/ros.h>
 
-#include <prbt_hardware_support/get_param.h>
+#include <pilz_utils/get_param.h>
 
-namespace prbt_hardware_support
+namespace pilz_utils
 {
 
 static const std::string DUMMY_PARAM_NAME{"dummy_parameter"};
@@ -77,7 +76,7 @@ TEST(GetParamTest, testAll)
   EXPECT_THROW(getParam<std::string>(nh, FALSE_PARAM_NAME), GetParamException);
 }
 
-} // namespace prbt_hardware_support
+} // namespace pilz_utils
 
 int main(int argc, char *argv[])
 {
