--- prbt_hardware_support/include/prbt_hardware_support/wait_for_timeouts.h
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2019 Pilz GmbH & Co. KG
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
-
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
-
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#ifndef WAIT_FOR_TIMEOUTS_H
-#define WAIT_FOR_TIMEOUTS_H
-
-namespace prbt_hardware_support
-{
-
-static constexpr double DEFAULT_RETRY_TIMEOUT {0.2};
-static constexpr double DEFAULT_MSG_OUTPUT_PERIOD {5.0};
-
-}
-
-#endif // WAIT_FOR_TIMEOUTS_H
