+++ prbt_hardware_support/package.xml
@@ -30,6 +30,7 @@
   <build_depend>tf2_ros</build_depend>
   <build_depend>urdf</build_depend>
   <build_depend>dynamic_reconfigure</build_depend>
+  <build_depend>pilz_utils</build_depend>
 
   <run_depend>roscpp</run_depend>
   <run_depend>message_runtime</run_depend>
