--- prbt_hardware_support/test/unittests/unittest_get_param.test
@@ -1,22 +1,0 @@
-<!--
-Copyright (c) 2019 Pilz GmbH & Co. KG
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
-<launch>
-  <test test-name="unittest_get_param" pkg="prbt_hardware_support" type="unittest_get_param">
-    <param name="dummy_parameter" value="parameter value"/>
-  </test>
-</launch>
